Development tools should now be installed via pipx
pipx was not able to properly get the flake8 plugins to work. So it
needs to stay in the Pipfile.

Because black was a dependency for "flake8-strict" and pipenv complains
about pre-release versions and black only has pre-release versions,
there was an error message. So it needs to stay in the Pipfile.